+++ app/src/main/java/com/android/hanple/ui/search/SearchTimeFragment.kt
@@ -1,12 +1,13 @@
-package com.android.hanple
+package com.android.hanple.ui.search
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import com.android.hanple.R
 
-class SearchFragment : Fragment() {
+class SearchTimeFragment : Fragment() {
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
