+++ app/src/main/java/com/android/hanple/ui/settings/SettingsFragment.kt
@@ -5,12 +5,11 @@ import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import android.widget.TextView
 import androidx.fragment.app.Fragment
 import androidx.lifecycle.ViewModelProvider
-import com.android.hanple.ArchiveActivity
-import com.android.hanple.MainActivity
+import com.android.hanple.ui.ArchiveActivity
 import com.android.hanple.databinding.FragmentSettingsBinding
+import com.android.hanple.viewmodel.SettingsViewModel
 
 class SettingsFragment : Fragment() {
     private var _binding: FragmentSettingsBinding? = null
