+++ app/src/main/java/com/android/hanple/viewmodel/ArchiveViewModel.kt
@@ -1,4 +1,4 @@
-package com.android.hanple
+package com.android.hanple.viewmodel
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
