+++ app/src/main/res/layout/activity_archive.xml
@@ -5,7 +5,7 @@
     android:id="@+id/main"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ArchiveActivity">
+    tools:context=".ui.ArchiveActivity">
 
     <ImageView
         android:id="@+id/iv_back_icon"
