+++ app/src/main/res/layout/activity_onboarding.xml
@@ -5,6 +5,6 @@
     android:id="@+id/main"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".OnboardingActivity">
+    tools:context=".ui.OnboardingActivity">
 
 </androidx.constraintlayout.widget.ConstraintLayout>
