+++ app/src/main/res/layout/fragment_search.xml
@@ -6,7 +6,7 @@
     android:background="@color/white"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".SearchFragment">
+    tools:context=".ui.search.SearchFragment">
 
     <com.google.android.material.card.MaterialCardView
         android:id="@+id/cd_search_place_you_want"
