+++ app/src/main/res/layout/fragment_search_schedule.xml
@@ -6,7 +6,7 @@
     android:background="@color/white"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".SearchScheduleFragment">
+    tools:context=".ui.search.SearchScheduleFragment">
 
     <com.google.android.material.card.MaterialCardView
         android:id="@+id/cd_search_schedule_enter_destination"
